refactor(core): name and bound the access-control two-phase read
Split packages/core/src/access/engine.ts into clearly named modules so the
two-phase read invariant is explicit in the code structure:

- access-filter.ts (Access Filter, phase 1, pre-query): buildIncludeWithAccessControl
- field-visibility.ts (Field Visibility, phase 2, post-query): filterReadableFields
- field-access.ts: canonical field-level evaluator (checkFieldAccess) +
  filterWritableFields, shared by read and write paths
- engine.ts: operation-level helpers (checkAccess, mergeFilters, type guards,
  getRelatedListConfig)

The repeated "check read access -> skip -> resolveOutput" sequence in the
post-query phase is extracted into a single unexported helper, removing the
duplication between the regular-field and virtual-field branches.

The access barrel re-exports every previously-exported name, so no importer
breaks. No behaviour change and no public API change.

Closes #394
