+++ packages/core/src/access/index.ts
@@ -11,14 +11,17 @@ export type {
   AugmentedFindUnique,
   FindManyQueryArgs,
 } from './types.js'
+// Operation-level access primitives and shared ref-parsing helper.
 export {
   checkAccess,
   mergeFilters,
-  checkFieldAccess,
-  filterReadableFields,
-  filterWritableFields,
   isBoolean,
   isPrismaFilter,
   getRelatedListConfig,
-  buildIncludeWithAccessControl,
 } from './engine.js'
+// Canonical field-level access evaluation (shared by read and write paths).
+export { checkFieldAccess, filterWritableFields } from './field-access.js'
+// Phase 1 — Access Filter (pre-query row/relation scoping).
+export { buildIncludeWithAccessControl } from './access-filter.js'
+// Phase 2 — Field Visibility (post-query field stripping + resolveOutput).
+export { filterReadableFields } from './field-visibility.js'
